pyproject.toml
[tool.poetry]
name = "tampermonkeyscripts"
version = "0.1.0"
description = ""
authors = ["YueLi <green07111@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = {extras = ["job-queue"], version = "^21.6"}
python-dotenv = "^1.0.1"
uv = "^0.4.20"
aiohttp = "^3.10.10"
pymongo = "^4.10.1"
tqdm = "^4.66.5"
requests = "^2.32.3"
numpy = "<2"
pydantic = "^2.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
